requirements.txt
django
djangorestframework
drf-spectacular
dj-database-url
gunicorn
whitenoise
psycopg2-binary
